requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.0.0
django-oauth-toolkit==1.7.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
psycopg-binary==3.0.8
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.13.3
zipp==3.7.0
